Adds fill = TRUE to rbindlist() al L1024
This aims to fix the error in binance_all_orders() caused by rbindlist() when trying to merge types of orders that differ on the number on elements. ("TAKE_PROFIT_LIMIT" for example). By adding the argument fill = TRUE to rbindlist() al L1024 the extra elements of such orders are added as extra columns in the data.frame,